Added all files

+++ src/com/company/Vector.java
@@ -1,22 +1,21 @@
 package com.company;
-
 public class Vector {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
